source/workspaces_app/requirements.txt
boto3==1.34.144
botocore==1.34.144
certifi==2024.07.04
charset-normalizer==2.1.1
idna==3.7
idna==3.7
jmespath==1.0.1
python-dateutil==2.8.2
requests==2.32.0
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
aws-lambda-powertools==2.39.1
